backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
mysqlclient>=2.2
python-dotenv>=1.0
django-cors-headers>=4.4
